tools: Drop gettext as a build dependency
It has not been a dependency since at least 4.13.  Remove its mandatory check
from ./configure.

Annotate the dependency in the CI dockerfiles which are used for 4.14 or
older.  Drop the dependency from Alpine and Ubuntu Focal which are new
containers in 4.15, as well as dropping from CirrusCI and TravisCI.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Ian Jackson <iwj@xenproject.org>
CC: Wei Liu <wl@xen.org>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Doug Goldstein <cardoe@cardoe.com>

CI Proofs of correctness.

  https://cirrus-ci.com/build/5554642114314240
  https://www.travis-ci.org/github/andyhhp/xen/builds/764514317 (x86 builds
  only.  ARM are permenantly broken now.)
  https://gitlab.com/xen-project/people/andyhhp/xen/-/pipelines/277055305 (using
  test Alpine and Focal containers rebuild without gettext).

For 4.15.  This is long overdue, and simplifies everyones life building 4.15.
If this doesn't go into 4.15, then we can't shrink the alpine/focal containers.