CI: Add zstd/etc libraries to build environments
zstd compression support is new in 4.15, but isn't currently tested anywhere
in CI.

bzip2/zlib was missing in some cases, so add those as well.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Ian Jackson <iwj@xenproject.org>
CC: Wei Liu <wl@xen.org>
CC: Doug Goldstein <cardoe@cardoe.com>

For 4.15.  Add build time testing for a currently untested feature.  This
should have been part of the series which included 8169f82049efb but I don't
think it really qualifies for a Fixes: tag.

RFC for two reasons.  First, this needs extending to other containers when
I've finished chasing package names, and the CI runs haven't completed yet.