x86: Use control flow typechecking where possible
Now all callees have been annotated, turn on typechecking to catch issues in
the future.

This extension isn't in a released version of GCC yet, so provide a container
to use with the extention included, and add it to CI.  RANDCONFIG is necessary
because some stubs for compiled-out subsystems are used as function pointer
targets.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Wei Liu <wl@xen.org>

v3:
 * Provide container.  Already pushed to Gitlab.