x86/boot: Don't mutate the module information mid-boot XXX broken
Right now, we have a loop in the middle of boot which performs:

  mod[i].mod_end -= mod[i].mod_start;
  mod[i].mod_start >>= PAGE_SHIFT;

This is straight obfuscation.  It creates a field named mod_end which is a
size, and a field intended to be a byte address being a frame address.

Furthermore, as bootstrap_map() only understands post-adjustment form, it
prevents loading microcode as early as we want.

Delete this part of the module loop, and adjust all subsequent logic in boot.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Roger Pau Monné <roger.pau@citrix.com>